+++ src/army/generic/index.ts
@@ -3,7 +3,7 @@ import GenericEndlessSpells from './endless_spells'
 import GenericScenery from './scenery'
 import GenericSpells from './spells'
 import GenericTriumphs from './triumphs'
-import RealmArtifacts from './realm_artifacts'
+import RealmscapeArtifacts from './realmscape_artifacts'
 import RealmscapeCommands from './realmscape_commands'
 import RealmscapeFeatures from './realmscape_features'
 import RealmscapeSpells from './realmscape_spells'
@@ -15,7 +15,7 @@ export {
   GenericScenery,
   GenericSpells,
   GenericTriumphs,
-  RealmArtifacts,
+  RealmscapeArtifacts,
   RealmscapeCommands,
   RealmscapeFeatures,
   RealmscapeSpells,
