+++ src/army/generic/realmscape_artifacts.ts
